Convert hydrated weight value to an integer

+++ Model/Order/HydrateOrderFromQuote.php
@@ -218,7 +218,7 @@ class HydrateOrderFromQuote implements HydrateOrderFromQuoteInterface
             $cartItem = $this->productFactory->create()->load($item['id']);
             $item['sku'] = $cartItem->getSku();
             $item['vendor'] = '';
-            $item['weight'] = ceil($item['weight']);
+            $item['weight'] = (int)ceil($item['weight']);
         }
 
         return $cartItems;
